+++ packages/frontend/src/components/Explorer/CustomMetricModal/FilterForm.tsx
@@ -8,7 +8,7 @@ import {
 } from '@lightdash/common';
 import { Button, Stack } from '@mantine/core';
 import { Dispatch, FC, SetStateAction, useCallback } from 'react';
-import { useExplorerContext } from '../../../providers/ExplorerProvider';
+import { useExplorerContext } from '../../../providers/ExplorerProvider/useExplorerContext';
 import FilterRuleForm from '../../common/Filters/FilterRuleForm';
 import { useFiltersContext } from '../../common/Filters/FiltersProvider';
 import { addFieldRefToFilterRule } from './utils';
