+++ packages/frontend/src/components/Explorer/ExploreTree/TableTree/Tree/TreeSingleNodeActions.tsx
@@ -24,7 +24,7 @@ import {
 } from '@tabler/icons-react';
 import { FC, useMemo } from 'react';
 import { useFilters } from '../../../../../hooks/useFilters';
-import { useExplorerContext } from '../../../../../providers/ExplorerProvider';
+import { useExplorerContext } from '../../../../../providers/ExplorerProvider/useExplorerContext';
 import { useTracking } from '../../../../../providers/TrackingProvider';
 import { EventName } from '../../../../../types/Events';
 import MantineIcon from '../../../../common/MantineIcon';
