+++ packages/frontend/src/components/Explorer/SaveChartButton/index.tsx
@@ -5,7 +5,7 @@ import { FC, useState } from 'react';
 import useToaster from '../../../hooks/toaster/useToaster';
 import { useAddVersionMutation } from '../../../hooks/useSavedQuery';
 import useSearchParams from '../../../hooks/useSearchParams';
-import { useExplorerContext } from '../../../providers/ExplorerProvider';
+import { useExplorerContext } from '../../../providers/ExplorerProvider/useExplorerContext';
 import MantineIcon from '../../common/MantineIcon';
 import ChartCreateModal from '../../common/modal/ChartCreateModal';
