+++ packages/frontend/src/components/Explorer/index.tsx
@@ -4,7 +4,7 @@ import { FC, memo } from 'react';
 import { useParams } from 'react-router-dom';
 import { useExplore } from '../../hooks/useExplore';
 import { useProjects } from '../../hooks/useProjects';
-import { useExplorerContext } from '../../providers/ExplorerProvider';
+import { useExplorerContext } from '../../providers/ExplorerProvider/useExplorerContext';
 import { CustomVisualizationProvider } from '../CustomVisualization';
 import { DrillDownModal } from '../MetricQueryData/DrillDownModal';
 import MetricQueryDataProvider from '../MetricQueryData/MetricQueryDataProvider';
