+++ packages/frontend/src/components/RefreshButton.tsx
@@ -10,7 +10,7 @@ import {
 import { useHotkeys, useOs } from '@mantine/hooks';
 import { IconPlayerPlay } from '@tabler/icons-react';
 import { FC, memo, useCallback } from 'react';
-import { useExplorerContext } from '../providers/ExplorerProvider';
+import { useExplorerContext } from '../providers/ExplorerProvider/useExplorerContext';
 import { useTracking } from '../providers/TrackingProvider';
 import { EventName } from '../types/Events';
 import MantineIcon from './common/MantineIcon';
